Added real game-logic in GameInstance. Added a couple more integration-tests, as well as started work on game-integration-tests and finished the first one. Lots of refactoring as well.

+++ GladiatorArenaServer/src/main/java/com/github/malow/gladiatorarena/server/comstructs/GetMyInfoResponse.java
@@ -7,7 +7,7 @@ public class GetMyInfoResponse extends Response
 {
   public String username;
   public Integer rating;
-  public Integer currentMatchId;
+  public Long currentMatchId;
   public boolean isSearchingForGame;
 
   public GetMyInfoResponse(boolean result, Player player)
@@ -19,7 +19,7 @@ public class GetMyInfoResponse extends Response
     this.isSearchingForGame = player.isSearchingForGame;
   }
 
-  public GetMyInfoResponse(boolean result, String username, Integer rating, Integer currentMatchId, boolean isSearchingForGame)
+  public GetMyInfoResponse(boolean result, String username, Integer rating, Long currentMatchId, boolean isSearchingForGame)
   {
     super(result);
     this.username = username;
